feat(import): finalizar importación legacy con feedback visual, limpieza de datos y soporte biométrico
- Migración de IDs de especie a 64 bits (long) para evitar desbordamientos.
- Implementación de limpieza automática de datos previos sincronizada con lances y producción.
- Soporte para campos de largo (total/estándar) y peso en la entidad ItemSubmuestra.
- Generación de migración EF Core para actualización de esquema de base de datos.
- Mejora de UI con mensajes de estado dinámicos (BusyMessage) y diálogos de confirmación premium.
- Corrección de errores de recursos XAML (estilos y colores faltantes).
- Actualización de pruebas de integración para el flujo de importación completo.

+++ OBSArrastre2026.App/Services/DbfExtractorService.cs
@@ -108,8 +108,8 @@ public sealed class DbfExtractorService : IDbfExtractorService
                 var espCode = GetDouble(reader, colMap, $"ESPECIE_{i}");
                 if (espCode > 0)
                 {
-                    c.Especies[(int)espCode] = GetDouble(reader, colMap, $"KG_{i}");
-                    c.DescartesPorEspecie[(int)espCode] = GetDouble(reader, colMap, $"DESCAR_{i}");
+                    c.Especies[(long)espCode] = GetDouble(reader, colMap, $"KG_{i}");
+                    c.DescartesPorEspecie[(long)espCode] = GetDouble(reader, colMap, $"DESCAR_{i}");
                 }
             }
             list.Add(c);
@@ -134,7 +134,7 @@ public sealed class DbfExtractorService : IDbfExtractorService
                 Lance = GetDouble(reader, colMap, "LANCE"),
                 Fecha = GetDateTime(reader, colMap, "FECHA") ?? DateTime.MinValue,
                 Especie = GetString(reader, colMap, "ESPECIE"),
-                CodEspec = (int)GetDouble(reader, colMap, "COD_ESPEC"),
+                CodEspec = (long)GetDouble(reader, colMap, "COD_ESPEC"),
                 Area = GetDouble(reader, colMap, "AREA"),
                 PrimTalla = (int)GetDouble(reader, colMap, "PRIM_TALLA"),
                 UltTalla = (int)GetDouble(reader, colMap, "ULT_TALLA"),
@@ -200,7 +200,7 @@ public sealed class DbfExtractorService : IDbfExtractorService
                 Marea = GetDouble(reader, colMap, "MAREA"),
                 Lance = GetDouble(reader, colMap, "LANCE"),
                 Fecha = GetDateTime(reader, colMap, "FECHA") ?? DateTime.MinValue,
-                CodEspecIE = (int)GetDouble(reader, colMap, "CODIGO"),
+                CodEspecIE = (long)GetDouble(reader, colMap, "CODIGO"),
             };
 
             for (int i = 1; i <= 70; i++)
